Move models files to separate folders
