+++ SKNIBot/Core/Commands/ModerationCommands/OnlineCommand.cs
@@ -10,7 +10,7 @@ using DSharpPlus.CommandsNext.Attributes;
 using DSharpPlus.Entities;
 using DSharpPlus.EventArgs;
 using SKNIBot.Core.Database;
-using SKNIBot.Core.Database.Models;
+using SKNIBot.Core.Database.Models.DynamicDB;
 using SKNIBot.Core.Helpers.Pagination;
 
 namespace SKNIBot.Core.Commands.ModerationCommands
