+++ SKNIBot/Core/Database/DynamicDBContext.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using SKNIBot.Core.Database.Logger;
-using SKNIBot.Core.Database.Models;
+using SKNIBot.Core.Database.Models.DynamicDB;
 
 namespace SKNIBot.Core.Database
 {
