+++ SKNIBot/Core/Services/EmojiCounterService.cs
@@ -1,6 +1,6 @@
 ﻿using DSharpPlus.Entities;
 using SKNIBot.Core.Database;
-using SKNIBot.Core.Database.Models;
+using SKNIBot.Core.Database.Models.DynamicDB;
 using System;
 using System.Collections.Generic;
 using System.Linq;
