+++ SKNIBot/Core/Services/WelcomeMessageService/WelcomeMessageService.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using SKNIBot.Core.Database;
-using SKNIBot.Core.Database.Models;
+using SKNIBot.Core.Database.Models.DynamicDB;
 using System;
 using System.Collections.Generic;
 using System.Linq;
